fix(crypto): hand the secret descriptor around as an int parsed at 31 bits
The unix duplicate converted the uintptr back to int, which the
integer-conversion analysis flags on the same grounds as before; parsing at
31 bits and passing an int keeps every conversion in range on every
platform. The daemon test harness now prints the daemon's stdout and stderr
when it fails to settle, so a runner-only failure can be read from CI.

+++ internal/crypto/passphrase_fd_windows.go
@@ -9,7 +9,7 @@ import (
 	"golang.org/x/sys/windows"
 )
 
-func duplicatePassphraseFD(fd uintptr) (*os.File, error) {
+func duplicatePassphraseFD(fd int) (*os.File, error) {
 	process := windows.CurrentProcess()
 	var duplicate windows.Handle
 	if err := windows.DuplicateHandle(
